"Add isSuspended and isWithdrawn fields to student object in adminUpdateStudent and writeExam functions"
This commit adds two new fields, isSuspended and isWithdrawn, to the student object. Also writing some logic to check if the student is suspended or withdrawn and throwing an error message if true. Adding back commented out code and removing the data returned in the request and replacing it with a string to inform the student to return back to check their results.