"Add populate methods to examResults controller and add answeredQuestions field to ExamResults schema"
This commit includes changes to the `examResults.js` controller in the `controller/academics` directory for the checkExamResults route. We are using a mongoose query to populate exam, class level, academic term, and academic year data of the exam the student is taking. This is based on the student ID that is generated and the ID of the request parameters. The return response has student data also returning to ensure the correct data is being passed. Will refactor to ensure data that is not necessary is removed. Needed to remove the logic to check if a user has already taken the exam for testing.