"Add pagination to getAllTeachersAdmin endpoint and remove deleteAdmin endpoint in adminRouter"
This commit adds pagination to the getAllTeachersAdmin endpoint in the teachersCtrl.js file by converting query strings to numbers then subtracting the page and multiplying the limit, we can determind how many should be skipped. Also created a start index and an end index. Query strings are working as expected.